Move ScaleDownBitmap to a new utils module
