+++ browser/screenshot/print_preview_extractor_internal.cc
@@ -20,6 +20,7 @@
 #include "base/memory/weak_ptr.h"
 #include "base/types/expected.h"
 #include "brave/browser/screenshot/print_preview_extractor.h"
+#include "brave/components/screenshot/core/browser/utils.h"
 #include "chrome/browser/pdf/pdf_pref_names.h"
 #include "chrome/browser/printing/print_compositor_util.h"
 #include "chrome/browser/printing/print_preview_data_service.h"
@@ -43,10 +44,7 @@
 #include "third_party/blink/public/common/associated_interfaces/associated_interface_provider.h"
 #include "third_party/skia/include/core/SkBitmap.h"
 #include "third_party/skia/include/core/SkCanvas.h"
-#include "third_party/skia/include/core/SkColor.h"
 #include "third_party/skia/include/core/SkImage.h"
-#include "third_party/skia/include/core/SkRect.h"
-#include "third_party/skia/include/core/SkSamplingOptions.h"
 #include "ui/gfx/codec/png_codec.h"
 
 #if BUILDFLAG(ENABLE_PDF)
@@ -63,47 +61,6 @@ using printing::mojom::PrintPreviewUI;
 using ImageCallback = screenshot::PrintPreviewExtractor::CaptureImagesCallback;
 
 namespace {
-
-// Scales `bitmap` so it fits within 1024x768 while preserving aspect ratio,
-// centering the result on a transparent canvas. Bitmaps already within those
-// bounds are returned unchanged. Kept local so the extractor doesn't pull in
-// ai_chat utilities.
-SkBitmap ScaleDownBitmap(const SkBitmap& bitmap) {
-  constexpr int kTargetWidth = 1024;
-  constexpr int kTargetHeight = 768;
-
-  if (bitmap.width() <= kTargetWidth && bitmap.height() <= kTargetHeight) {
-    return bitmap;
-  }
-
-  SkBitmap scaled_bitmap;
-  scaled_bitmap.allocN32Pixels(kTargetWidth, kTargetHeight);
-
-  SkCanvas canvas(scaled_bitmap);
-  canvas.clear(SK_ColorTRANSPARENT);
-
-  SkSamplingOptions sampling_options(SkFilterMode::kLinear,
-                                     SkMipmapMode::kLinear);
-
-  float src_aspect = static_cast<float>(bitmap.width()) / bitmap.height();
-  float dst_aspect = static_cast<float>(kTargetWidth) / kTargetHeight;
-
-  SkRect dst_rect;
-  if (src_aspect > dst_aspect) {
-    // Source is wider — fit to width.
-    float scaled_height = kTargetWidth / src_aspect;
-    float y_offset = (kTargetHeight - scaled_height) / 2;
-    dst_rect = SkRect::MakeXYWH(0, y_offset, kTargetWidth, scaled_height);
-  } else {
-    // Source is taller — fit to height.
-    float scaled_width = kTargetHeight * src_aspect;
-    float x_offset = (kTargetWidth - scaled_width) / 2;
-    dst_rect = SkRect::MakeXYWH(x_offset, 0, scaled_width, kTargetHeight);
-  }
-
-  canvas.drawImageRect(bitmap.asImage(), dst_rect, sampling_options);
-  return scaled_bitmap;
-}
 
 // chrome/browser/printing/print_view_manager_common.cc
 // Pick the right RenderFrameHost based on the WebContents.
