+++ browser/ui/webui/ai_chat/BUILD.gn
@@ -37,6 +37,7 @@ source_set("ai_chat") {
     "//brave/components/constants",
     "//brave/components/resources:static_resources_grit",
     "//brave/components/resources:strings_grit",
+    "//brave/components/screenshot/core/browser",
     "//chrome/browser/bookmarks",
     "//chrome/browser/favicon",
     "//chrome/browser/history",
