+++ components/ai_chat/content/browser/BUILD.gn
@@ -49,6 +49,7 @@ static_library("browser") {
     "//brave/components/brave_stats/browser",
     "//brave/components/constants",
     "//brave/components/resources:strings_grit",
+    "//brave/components/screenshot/core/browser",
     "//brave/components/text_recognition/common/buildflags",
     "//components/favicon/content",
     "//components/favicon/core",
