+++ components/ai_chat/core/browser/utils.cc
@@ -223,49 +223,6 @@ EngineConsumer::GenerationDataCallback BindParseRewriteReceivedData(
       std::move(callback));
 }
 
-SkBitmap ScaleDownBitmap(const SkBitmap& bitmap) {
-  constexpr int kTargetWidth = 1024;
-  constexpr int kTargetHeight = 768;
-
-  // Don't need to scale if dimensions are already smaller than target
-  // dimensions
-  if (bitmap.width() <= kTargetWidth && bitmap.height() <= kTargetHeight) {
-    return bitmap;
-  }
-
-  SkBitmap scaled_bitmap;
-  scaled_bitmap.allocN32Pixels(kTargetWidth, kTargetHeight);
-
-  SkCanvas canvas(scaled_bitmap);
-  canvas.clear(SK_ColorTRANSPARENT);
-
-  // Use high-quality scaling options
-  SkSamplingOptions sampling_options(SkFilterMode::kLinear,
-                                     SkMipmapMode::kLinear);
-
-  // Maintain aspect ratio while fitting within target dimensions
-  float src_aspect = static_cast<float>(bitmap.width()) / bitmap.height();
-  float dst_aspect = static_cast<float>(kTargetWidth) / kTargetHeight;
-
-  SkRect dst_rect;
-  if (src_aspect > dst_aspect) {
-    // Source is wider - fit to width
-    float scaled_height = kTargetWidth / src_aspect;
-    float y_offset = (kTargetHeight - scaled_height) / 2;
-    dst_rect = SkRect::MakeXYWH(0, y_offset, kTargetWidth, scaled_height);
-  } else {
-    // Source is taller - fit to height
-    float scaled_width = kTargetHeight * src_aspect;
-    float x_offset = (kTargetWidth - scaled_width) / 2;
-    dst_rect = SkRect::MakeXYWH(x_offset, 0, scaled_width, kTargetHeight);
-  }
-
-  // Draw scaled bitmap with high-quality sampling
-  canvas.drawImageRect(bitmap.asImage(), dst_rect, sampling_options);
-
-  return scaled_bitmap;
-}
-
 GURL GetEndpointUrl(bool premium, const std::string& path) {
   CHECK(!path.starts_with("/"));
